edit : 주가지수 type 변경 - int -> double

+++ Back-end/core-module/src/main/java/com/ico/core/dto/InvestReqDto.java
@@ -16,14 +16,14 @@ import javax.validation.constraints.NotNull;
 @NoArgsConstructor
 public class InvestReqDto {
     @NotNull(message = "714")
-    private Integer price;
+    private Double price;
 
     @NotNull(message = "715")
     @Min(value = 0, message = "716")
     private Integer amount;
 
     @Builder
-    public InvestReqDto(Integer price, Integer amount) {
+    public InvestReqDto(Double price, Integer amount) {
         this.price = price;
         this.amount = amount;
     }
